backend/requirements.txt
# MongoDB async driver
pymongo>=4.9,<5.0
# JWT auth
python-jose[cryptography]
fastapi
uvicorn
python-dotenv
langchain-groq
python-dotenv
google-generativeai
groq
uagents
reportlab
